Improve k8s test
