+++ sdk/identity/identity/test/manual-integration/Kubernetes/Dockerfile
@@ -1,0 +1,19 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+ARG NODE_VERSION=10
+
+# docker can't tell when the repo has changed and will therefore cache this layer
+FROM alpine/git as repo
+RUN git clone https://github.com/jonathandturner/azure-sdk-for-js --single-branch --branch manual_integration_test_1 --depth 1 /azure-sdk-for-js
+
+FROM node:${NODE_VERSION}-slim
+
+COPY --from=repo /azure-sdk-for-js/sdk/identity /sdk/identity
+COPY --from=repo /azure-sdk-for-js/sdk/core /sdk/core
+COPY --from=repo /azure-sdk-for-js/sdk/keyvault/keyvault-secrets /sdk/keyvault/keyvault-secrets
+
+WORKDIR /sdk/identity/identity/tests/manual_integration/kubernetes
+RUN node install
+CMD ["node", "index"]
